Refactor the way built-in commands are overridden.
You still define a eg build() in the Dofile. But now it must be
decorated with @command so that the syntax checker will ignore it.

It actually gets put into the modules 'build' property, so it is
still available through this_module() and when importing the module
as an object.